refactor: finish consolidating the send flow into the transfer domain
