+++ src/features/transfer/screens/SendSheet.tsx
@@ -39,11 +39,8 @@ import useAccountSettings from '@/hooks/useAccountSettings';
 import useCoinListEditOptions from '@/hooks/useCoinListEditOptions';
 import useColorForAsset from '@/hooks/useColorForAsset';
 import useContacts from '@/hooks/useContacts';
-import useMaxInputBalance from '@/hooks/useMaxInputBalance';
 import { usePersistentDominantColorFromImage } from '@/hooks/usePersistentDominantColorFromImage';
 import usePrevious from '@/hooks/usePrevious';
-import useSendableUniqueTokens from '@/hooks/useSendableUniqueTokens';
-import useSendSheetInputRefs from '@/hooks/useSendSheetInputRefs';
 import useUserAccounts from '@/hooks/useUserAccounts';
 import { logger, RainbowError } from '@/logger';
 import { useNavigation } from '@/navigation/Navigation';
@@ -51,7 +48,6 @@ import Routes from '@/navigation/routesNames';
 import { type RootStackParamList } from '@/navigation/types';
 import { type Contact } from '@/redux/contacts';
 import { rainbowTokenList } from '@/references/rainbow-token-list';
-import { getDefaultCheckboxes } from '@/screens/SendConfirmationSheet';
 import { useUserAssetsStore } from '@/state/assets/userAssets';
 import { useBackendNetworksStore } from '@/state/backendNetworks/backendNetworks';
 import { ChainId } from '@/state/backendNetworks/types';
@@ -67,10 +63,14 @@ import SendAssetForm from '../components/SendAssetForm';
 import SendAssetList from '../components/SendAssetList';
 import SendContactList from '../components/SendContactList';
 import SendHeader from '../components/SendHeader';
+import useMaxInputBalance from '../hooks/useMaxInputBalance';
+import useSendableUniqueTokens from '../hooks/useSendableUniqueTokens';
 import { useSendChainState } from '../hooks/useSendChainState';
+import useSendSheetInputRefs from '../hooks/useSendSheetInputRefs';
 import { useSendSubmit } from '../hooks/useSendSubmit';
 import { useSponsoredSendPreparation } from '../hooks/useSponsoredSendPreparation';
 import { getSendSubmitButtonState } from '../utils/sendSheetUtils';
+import { getDefaultCheckboxes } from './SendConfirmationSheet';
 
 const sheetHeight = deviceUtils.dimensions.height - (Platform.OS === 'android' ? 30 : 10);
